pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-transcriber"
version = "0.1.0"
description = "Audio extraction and transcription utility"
requires-python = ">=3.8,<3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
    "numpy>=1.21.0",
    "tqdm>=4.65.0",
    "click",
    "pyaudioop"
]

[tool.setuptools.packages.find]
include = ["audio_transcriber*"]

[project.scripts]
audio-transcriber = "audio_transcriber.audio_processor:main"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 